requirements.txt
# Core ML and Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.9.0
transformers>=4.0.0
sentence-transformers>=0.4.0

# Text Processing
nltk>=3.6
spacy>=3.0.0
textblob>=0.15.0

# Graph and Network Analysis
networkx>=2.6
pygraphviz>=0.9

# Knowledge Graph and Ontology
rdflib>=6.0.0
owlready2>=0.24

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.60.0
joblib>=1.0.0
python-dotenv>=0.19.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=21.0.0
flake8>=4.0.0